\name{NB-package}

\alias{NB-package}

\alias{NB}

\docType{package}

\title{Estimating effective population size from genetic data}

\description{
The concept of effective population size (\eqn{N_e}) is an inportant concept in population genetics. It tells the "effective" number of individuals within a population, that is, those who are able to reproduce. 
The allele frequencies in a closed population change over time as known as genetic drift, and the magnitude of this fluctuation is directly related to \eqn{N_e}. 
This package aims to estimate this quantity from genetic samples collected over multiple time points. 
}

\details{
\tabular{ll}{
Package: \tab NB\cr
Type: \tab Package\cr
Version: \tab 0.9\cr
Date: \tab 2014-10-03\cr
License: \tab GPL (>= 2) \cr
}
}

\author{Tin-Yu Hui}

\references{Please email me for referencing or citation. }

\keyword{Effective population size. Maximum Likelihood Estimation. }
