\name{NB.example.dataset}

\alias{NB.example.dataset}

\title{Generate sample dataset}

\description{To create a sample dataset in plain text format to illustrate the use of this package. }

\usage{NB.example.dataset()}

\details{No arguments required. Just run the function. }

\note{A plain text (*txt) file will be created in your current working directory. Please make sure that you have the permission to write onto the working directory. }

\seealso{
 \code{\link{NB.estimator}}
}

\examples{
## CREATE A SAMPLE DATASET IN PLAIN TEXT FORM
NB.example.dataset()

## CHECK YOUR WORKING DIRECTORY!

}
